requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.5
colorama>=0.4.6
tqdm>=4.66.1
python-dotenv>=1.0.0
pyfiglet>=0.8.post1
tabulate>=0.9.0
requests>=2.31.0
prompt_toolkit>=3.0.39